M2-5078: moved shared logic from TansferOwnership in the hook
