+++ src/modules/Dashboard/features/Applet/Popups/TransferOwnershipPopup/TransferOwnershipPopup.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Modal } from 'shared/components';
@@ -6,7 +5,7 @@ import { popups, applet } from 'redux/modules';
 import { useAppDispatch } from 'redux/store';
 import { TransferOwnership } from 'modules/Dashboard/features/Applet/TransferOwnership';
 import { StyledModalWrapper } from 'shared/styles/styledComponents';
-import { Mixpanel } from 'shared/utils';
+import { useTransferOwnership } from 'shared/hooks/useTransferOwnership';
 
 import { SuccessTransferOwnershipPopup } from '../SuccessTransferOwnershipPopup';
 
@@ -16,11 +15,14 @@ export const TransferOwnershipPopup = () => {
   const { transferOwnershipPopupVisible, applet: appletData } = popups.useData();
   const { result } = applet.useAppletData() || {};
   const currentApplet = appletData || result;
-
-  const [transferOwnershipSuccessVisible, setTransferOwnershipSuccessVisible] = useState(false);
-
-  const [isSubmitted, setIsSubmitted] = useState(false);
-  const [emailTransfered, setEmailTransfered] = useState('');
+  const {
+    transferOwnershipSuccessVisible,
+    isSubmitted,
+    setIsSubmitted,
+    emailTransfered,
+    setEmailTransfered,
+    handleSubmit,
+  } = useTransferOwnership();
 
   const transferOwnershipPopupClose = () => {
     dispatch(
@@ -31,17 +33,6 @@ export const TransferOwnershipPopup = () => {
       }),
     );
   };
-
-  const handleSubmit = () => {
-    setIsSubmitted(true);
-  };
-
-  useEffect(() => {
-    if (!emailTransfered) return;
-
-    setTransferOwnershipSuccessVisible(true);
-    Mixpanel.track('Invitation sent successfully');
-  }, [emailTransfered]);
 
   return (
     <>
