+++ src/shared/features/AppletSettings/TransferOwnershipSetting/TransferOwnershipSetting.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Box } from '@mui/material';
 
@@ -6,7 +6,7 @@ import { applet } from 'shared/state';
 import { TransferOwnership } from 'modules/Dashboard/features/Applet/TransferOwnership';
 import { SuccessTransferOwnershipPopup } from 'modules/Dashboard/features/Applet/Popups';
 import { TransferOwnershipRef } from 'modules/Dashboard/features/Applet/TransferOwnership/TransferOwnership.types';
-import { Mixpanel } from 'shared/utils';
+import { useTransferOwnership } from 'shared/hooks/useTransferOwnership';
 
 import { StyledTransferOwnershipForm } from './TransferOwnershipSetting.styles';
 import { StyledAppletSettingsButton } from '../AppletSettings.styles';
@@ -14,26 +14,24 @@ import { StyledAppletSettingsButton } from '../AppletSettings.styles';
 export const TransferOwnershipSetting = () => {
   const { t } = useTranslation('app');
   const { result: appletData } = applet.useAppletData() ?? {};
-
-  const [isSubmitted, setIsSubmitted] = useState(false);
-  const [emailTransfered, setEmailTransfered] = useState('');
-  const [transferOwnershipPopupVisible, setTransferOwnershipPopupVisible] = useState(false);
+  const {
+    transferOwnershipSuccessVisible,
+    setTransferOwnershipSuccessVisible,
+    isSubmitted,
+    setIsSubmitted,
+    emailTransfered,
+    setEmailTransfered,
+    handleSubmit,
+  } = useTransferOwnership();
   const transferOwnershipRef = useRef<TransferOwnershipRef | null>(null);
 
   const dataTestid = 'applet-settings-transfer-ownership';
 
   const handleSuccessPopupClose = () => {
-    setTransferOwnershipPopupVisible(false);
+    setTransferOwnershipSuccessVisible(false);
     setEmailTransfered('');
     transferOwnershipRef.current?.resetEmail();
   };
-
-  useEffect(() => {
-    if (!emailTransfered) return;
-
-    setTransferOwnershipPopupVisible(true);
-    Mixpanel.track('Invitation sent successfully');
-  }, [emailTransfered]);
 
   return (
     <>
@@ -51,7 +49,7 @@ export const TransferOwnershipSetting = () => {
       <Box sx={{ width: 'fit-content' }}>
         <StyledAppletSettingsButton
           variant="outlined"
-          onClick={() => setIsSubmitted(true)}
+          onClick={handleSubmit}
           data-testid={`${dataTestid}-confirm`}
         >
           {t('confirm')}
@@ -59,7 +57,7 @@ export const TransferOwnershipSetting = () => {
       </Box>
       <SuccessTransferOwnershipPopup
         email={emailTransfered}
-        transferOwnershipPopupVisible={transferOwnershipPopupVisible}
+        transferOwnershipPopupVisible={transferOwnershipSuccessVisible}
         closeTransferOwnershipPopup={handleSuccessPopupClose}
         data-testid={`${dataTestid}-success-popup`}
       />
