+++ src/shared/hooks/index.ts
@@ -20,3 +20,4 @@ export * from './useNetwork';
 export * from './useCheckIfAppletHasNotFoundError';
 export * from './useRespondentLabel';
 export * from './useWindowFocus';
+export * from './useTransferOwnership';
